requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
sqlalchemy==2.0.40
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
